+++ packages/expo-updates/plugin/src/withUpdates.ts
@@ -1,8 +1,5 @@
 import { getAccountUsername } from '@expo/config';
-import { ConfigPlugin, createRunOncePlugin } from '@expo/config-plugins';
-
-import { withUpdatesAndroid } from './withUpdatesAndroid';
-import { withUpdatesIOS } from './withUpdatesIOS';
+import { ConfigPlugin, createRunOncePlugin, AndroidConfig, IOSConfig } from '@expo/config-plugins';
 
 const pkg = require('expo-updates/package.json');
 
@@ -10,8 +7,8 @@ const withUpdates: ConfigPlugin<{ expoUsername?: string } | void> = (config, pro
   // The username will be passed from the CLI when the plugin is automatically used.
   const expoUsername = (props || {}).expoUsername ?? getAccountUsername(config);
 
-  config = withUpdatesAndroid(config, { expoUsername });
-  config = withUpdatesIOS(config, { expoUsername });
+  config = AndroidConfig.Updates.withUpdates(config, { expoUsername });
+  config = IOSConfig.Updates.withUpdates(config, { expoUsername });
   return config;
 };